Moved out speed, direction and positioning
It's small steps trying to untangle all related state. Some of the
tangles are required, some intentional and some are to fix bugs.
It's getting tough because I am zeroing in on the intersection
between drag movement and manual movement. On top of that, the
structure is hidden because all moving parts live in the same scope.
If the author would ever read this, my major points of advise would be
to try and keep the following rules for functions:

- Either return a value or perform side effects, never both
- Keep everything at the same level of abstraction