Rename hook to fireEvent
Also fixed a sneaky bug. Caused by passing `pubfuncs` in its current
state. All `pubfuncs` methods interact with `slider.current`. If an
event is fired during the construction of `KeenSlider` a strange
situation occurs. The event fires from a `KeenSlider` that is not
yet stored in `slider.current`, this means that any calls from an
event handler would target the wrong slider.

Introducing the `mount` function solves this problem. We could
consider passing the `fireEvent` function into the `mount` function
to prevent accidental calls to `fireEvent` from the contructor.