Cleanup
Many different changes. Mainly focussed on formatting and naming. Also a
few steps in restructuring and typing. The main reason for this type of
work is to improve readability. I try to take a step back and see how I
can improve the code reading experience. As a side effect it might
inspire me into new directions of structure.