Move last dynamic options to the outside
At this point all option related 'moving parts' (dynamic options) have
been separated. It will now be easier to determine when the behavior
of the slider will change. It also opens interesting avenues for
extracting behavior. This in turn will allow in the future for more
focused behavior and separation of concerns.