Working towards the concept of strategy
At the moment I do not think this is a great improvement, it however
extracts some of the logic from the slider. The slider is assumes slides
of a fixed size and requires some calculation to determine that size.
There are however use cases where the slide size is known or the slide
size varies that can forego these calculations. By moving this into a
strategy we allow ourselves to use different strategies for positioning.