Big feature removal
Started a big cleanup. We are removing a lot of the complication by
reducing the amount of features. Most features that are removed
could relatively easy be added when using a framework like React
and the provided events.

Some might be re-introduced after discussing things like the original
CSS file.